2022 day20 and 21
