--- advent_of_code/aoc/2022/inputs/day20_sample.txt
@@ -1,7 +1,0 @@
-1
-2
--3
-3
--2
-0
-4
